refactor: update swap to use templating
