--- omnichain/swap/contracts/TestContracts.sol
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity 0.8.7;
-
-import "@zetachain/toolkit/contracts/BytesHelperLib.sol";
-import "@zetachain/toolkit/contracts/TestSystemContract.sol";
-import "@zetachain/toolkit/contracts/TestZRC20.sol";
