+++ omnichain/swap/package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint:fix": "npx eslint . --ext .js,.ts --fix",
+    "lint": "npx eslint . --ext .js,.ts"
   },
   "keywords": [],
   "author": "",
@@ -12,7 +14,7 @@
   "devDependencies": {
     "@ethersproject/abi": "^5.4.7",
     "@ethersproject/providers": "^5.4.7",
-    "@inquirer/select": "^1.1.7",
+    "@inquirer/select": "^1.1.3",
     "@nomicfoundation/hardhat-chai-matchers": "^1.0.0",
     "@nomicfoundation/hardhat-network-helpers": "^1.0.0",
     "@nomicfoundation/hardhat-toolbox": "^2.0.0",
@@ -23,23 +25,29 @@
     "@types/chai": "^4.2.0",
     "@types/mocha": ">=9.1.0",
     "@types/node": ">=12.0.0",
-    "@uniswap/v2-core": "^1.0.1",
-    "@uniswap/v2-periphery": "^1.1.0-beta.0",
-    "@zetachain/addresses": "^0.0.10",
-    "@zetachain/addresses-tools": "^0.0.8",
-    "@zetachain/faucet-cli": "^2.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "@typescript-eslint/parser": "^5.59.9",
+    "@zetachain/faucet-cli": "^2.0.1-beta.2",
     "@zetachain/interfaces": "^0.0.1",
     "@zetachain/networks": "^2.3.0-athens3",
-    "@zetachain/protocol-contracts": "^1.0.0-athens3",
-    "@zetachain/toolkit": "^1.0.2-athens3",
-    "@zetachain/zevm-example-contracts": "^0.0.1",
+    "@zetachain/protocol-contracts": "^1.0.1-athens3",
+    "@zetachain/toolkit": "^1.0.8-athens3",
     "axios": "^1.3.6",
     "chai": "^4.2.0",
     "dotenv": "^16.0.3",
     "envfile": "^6.18.0",
+    "eslint": "^8.42.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-sort-keys-fix": "^1.1.2",
+    "eslint-plugin-typescript-sort-keys": "^2.3.0",
     "ethers": "^5.4.7",
     "hardhat": "^2.14.0",
     "hardhat-gas-reporter": "^1.0.8",
+    "prettier": "^2.8.8",
     "solidity-coverage": "^0.8.0",
     "ts-node": ">=8.0.0",
     "typechain": "^8.1.0",
