+++ omnichain/swap/tasks/deploy.ts
@@ -1,8 +1,6 @@
+import { getAddress } from "@zetachain/protocol-contracts";
 import { task } from "hardhat/config";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
-import { getAddress } from "@zetachain/protocol-contracts";
-
-const contractName = "ZetaSwapV2";
 
 const main = async (args: any, hre: HardhatRuntimeEnvironment) => {
   if (hre.network.name !== "zeta_testnet") {
@@ -14,10 +12,10 @@ const main = async (args: any, hre: HardhatRuntimeEnvironment) => {
   const [signer] = await hre.ethers.getSigners();
   console.log(`🔑 Using account: ${signer.address}\n`);
 
-  const SYSTEM_CONTRACT = getAddress("systemContract", hre.network.name);
+  const systemContract = getAddress("systemContract", "zeta_testnet");
 
-  const factory = await hre.ethers.getContractFactory(contractName);
-  const contract = await factory.deploy(SYSTEM_CONTRACT);
+  const factory = await hre.ethers.getContractFactory("Swap");
+  const contract = await factory.deploy(systemContract);
   await contract.deployed();
 
   console.log(`🚀 Successfully deployed contract on ZetaChain.
